shear: 🐑 initial commit
this project is still a work-in-progress, but is now far enough along so
as to be ready to be checked into version control.

this is `shear`, a library for trimming excess contents from things.

right now, this contains generic, core infrastructure providing a state
machine to iterate across an iterator of some kind, and to potentially
truncate the items of the sequence. these core facilities are used to
provide an interface for trimming strings and leaving an ellipses at the
end.