str: `trim_to_length()` accepts a configurable `Ellipsis`
this makes the precise sequence of characters that are used to truncate
a string configurable, via a public `Ellipsis` trait. this allows
callers to provide an arbitrary static string.

two ellipses are provided to start, `...` and `…`. a new
`str::LimitedExt` trait is included, which provides these `E` receivers
on behalf of the caller.